清理******中编译告警
# message auto-generated for no-merge-commit merge:
merge warning_remove into master

【开发自提】【********】 清理******中存在的编译告警

Created-by: insightDeveloper

Match-id-c3c1510a67eb7fa4c07fda0bbdf4057588bb1523

+++ server/src/modules/source/parser/SourceFileParser.cpp
@@ -78,7 +78,7 @@ bool SourceFileParser::Parse(const std::vector<std::string> &filePaths, const st
         dataType = static_cast<int>(dataType);
         paddingLength = static_cast<int>(paddingLength);
         if (dataType == static_cast<int>(DataTypeEnum::SOURCE)) {
-            if (INT64_MAX - filePathLen < dataSize) {
+            if (static_cast<uint64_t>(INT64_MAX - filePathLen) < dataSize) {
                 // 溢出防护
                 ServerLog::Error("Source code data block in selected file is invalid which data size is :", dataSize);
                 return false;
